Upd: use cassandra 3.0.28

+++ central/src/test/java/org/glowroot/central/repo/ConfigRepositoryIT.java
@@ -101,6 +101,9 @@ public class ConfigRepositoryIT {
 
     @AfterAll
     public static void tearDown() throws Exception {
+        if (!SharedSetupRunListener.isStarted()) {
+            return;
+        }
         asyncExecutor.shutdown();
         // remove bad data so other tests don't have issue
         session.updateSchemaWithRetry("drop table if exists agent_config");
